openaction/requirements.txt
PIL
-e git://github.com/revsys/django-friendship.git#egg=django-friendship
django-social-auth
-e git://github.com/carljm/django-model-utils.git#egg=django-model-utils
django-registration
-e git://github.com/brosner/django-notification.git#egg=django-notification
redis
